+++ lib/routes/app_pages.dart
@@ -4,14 +4,15 @@ import 'package:team_a_project/pages/condition_page.dart';
 import 'package:team_a_project/pages/decide_page.dart';
 import 'package:team_a_project/pages/home.dart';
 import 'package:team_a_project/pages/riku_app.dart';
-import 'package:team_a_project/pages/yoshi_app.dart';
+import 'package:team_a_project/pages/yoshi_app/yoshi_app.dart';
+import 'package:team_a_project/pages/yoshi_app/task_add.dart';
 import 'package:team_a_project/routes/app_routes.dart';
 
 class AppPages {
   static final List<GetPage<dynamic>> list = <GetPage<dynamic>>[
     GetPage(
       name: AppRoutes.home,
-      page: () =>  MyHomePage(),
+      page: () => MyHomePage(),
     ),
     GetPage(
       name: AppRoutes.yoshiApp,
@@ -29,5 +30,7 @@ class AppPages {
       name: AppRoutes.decidePage,
       page: () => const DecidePage(),
     ),
+    GetPage(
+        name: AppRoutes.yoshiAppAddTask, page: () => const YoshiTaskAddPage())
   ];
 }
