+++ lib/routes/app_routes.dart
@@ -2,6 +2,7 @@ class AppRoutes {
   static const initial = "/";
   static const home = "/home";
   static const yoshiApp = "/yoshiApp";
+  static const yoshiAppAddTask = "/YoshiTaskAddPage";
   static const rikuApp = "/rikuApp";
   static const conditionPage = "/conditionPage";
   static const decidePage = "/decidePage";
